openai implements llm interface
but with panics, will be fixed in upcoming pr
