+++ models.go
@@ -6,42 +6,42 @@ type Model struct {
 }
 
 var ModelO3Mini Model = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "o1-mini",
 }
 
 var ModelO1 Model = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "o1",
 }
 
 var ModelO1Mini Model = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "o1-mini",
 }
 
 var ModelO1Preview Model = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "o1-preview",
 }
 
 var ModelGPT4 Model = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "gpt-4",
 }
 
 var ModelGPT4Turbo Model = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "gpt-4-turbo",
 }
 
 var ModelGPT4O = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "gpt-4o",
 }
 
 var ModelGPT4OMini Model = Model{
-	Provider: ProviderOpenAI,
+	Provider: Openai{},
 	Name:     "gpt-4o-mini",
 }
